requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cssbeautifier==1.15.1
Django==5.1.3
django-celery-beat==2.7.0
django-timezone-field==7.0
djangorestframework==3.15.2
djlint==1.36.3
EditorConfig==0.12.4
jsbeautifier==1.15.1
json5==0.10.0
kombu==5.4.2
pathspec==0.12.1
pre-commit==4.1.0
prompt_toolkit==3.0.48
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
regex==2024.11.6
six==1.16.0
sqlparse==0.5.2
tqdm==4.67.1
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
